+++ packages/hubble-db/src/models/FarmUserStateEntity.ts
@@ -1,7 +1,5 @@
-import Decimal from 'decimal.js';
-
 export interface FarmUserStateEntity {
-  id: Decimal;
+  id: BigInt;
   raw_json: string;
   created_on: Date;
   owner_id: number;
